Feat: Add an extra authorization to the endpoint for preventing unauthorized request for deletion

+++ Queueomatic/Server/Endpoints/Participant/Delete/DeleteParticipantEndpoint.cs
@@ -1,4 +1,5 @@
 ﻿using FastEndpoints;
+using Queueomatic.DataAccess.UnitOfWork;
 using Queueomatic.Server.Services.ParticipantService;
 
 namespace Queueomatic.Server.Endpoints.Participant.Delete;
@@ -6,10 +7,11 @@ namespace Queueomatic.Server.Endpoints.Participant.Delete;
 public class DeleteParticipantEndpoint: Endpoint<DeleteParticipantRequest>
 {
     private readonly IParticipantService _participantService;
-
-    public DeleteParticipantEndpoint(IParticipantService participantService)
+    private readonly IUnitOfWork _unitOfWork;
+    public DeleteParticipantEndpoint(IParticipantService participantService, IUnitOfWork unitOfWork)
     {
         _participantService = participantService;
+        _unitOfWork = unitOfWork;
     }
 
     public override void Configure()
@@ -20,7 +22,38 @@ public class DeleteParticipantEndpoint: Endpoint<DeleteParticipantRequest>
 
     public override async Task HandleAsync(DeleteParticipantRequest req, CancellationToken ct)
     {
+        if (req.ParticipantId != null && 
+            req.ParticipantId != req.Id)
+        {
+            await SendUnauthorizedAsync();
+            return;
+        }
+
+        if (!string.IsNullOrEmpty(req.UserId) && 
+            !await IsUserOwnerOfTheRoom(req))
+        {
+            await SendUnauthorizedAsync();
+            return;
+        }
+        
         await _participantService.DeleteOneAsync(req.Id);
         await SendNoContentAsync();
+    }
+
+    private async Task<bool> IsUserOwnerOfTheRoom(DeleteParticipantRequest req)
+    {
+        var user = await _unitOfWork.UserRepository.GetAsync(req.UserId);
+        if (user == null) return false;
+
+        var userRooms = user.Rooms.ToList();
+        
+        var isUserTheOwner =
+            userRooms.FirstOrDefault(r =>
+                r.Participators.FirstOrDefault(p =>
+                    p.Id == req.Id) != null);
+
+        if (isUserTheOwner == null) return false;
+
+        return true;
     }
 }
